Change product information and test to lombok
